requirements.txt
streamlit
opencv-python-headless
numpy
mediapipe
Pillow
